relations resolved and repositories are ready for implementation

+++ DigitalShoes.Domain/Entities/ApplicationUser.cs
@@ -3,15 +3,14 @@
 namespace DigitalShoes.Domain.Entities
 {    
     public class ApplicationUser:IdentityUser<int>
-    {        
-        public string Name { get; set; }
-        //public Image Image { get; set; }        
+    {
+        public string Name { get; set; }        
         public string ProfileImageUrl { get; set; }
         public string ProfileImageLocalPath { get; set; }
         public Cart Cart { get; set; }                
         public Wishlist Wishlist { get; set; }
         public List<Review> Reviews { get; set; }
         public List<Payment> Payments { get; set; }
-        public int Balance { get; set; }
+        public decimal Balance { get; set; }
     }
 }
